+++ packages/core/internal.d.ts
@@ -3,3 +3,4 @@
 export { $NODE, $OWN_KEYS, $PROXY, $RAW, $VERSION } from "./src/core";
 export { setProperty, deleteProperty, bumpOwnKeysSignal, bumpVersion } from "./src/write";
 export { profileSignalWrite } from "./src/profiler";
+export { startBatch, endBatch, getCurrentSub, setCurrentSub } from "alien-signals";